Complete design puzzle for 1st chapter intro to design pattern: Strategy pattern.
The organization of the Design Puzzle came out nicely.

All of the characters including both the abstract and its
implemenntation of characters are put into the Characters folder.

And the weapon behavior interface and all of its implementations are
 orgnized into the WeaponBehaviours folder.

Both of the above one level deep. This is great. Having to deal with the
many level of indentation of classes/folders has been one great code
smell I've to realise. By not having that problem here, indicates that so far
the strategy pattern is good one to keep aware of.

I'm not sure yet how I could use it in a real project.

$ php src/_01_TheStrategyPattern/DesignPuzzle/MiniActionAdventureGameSimulator.php
Troll swings axe.
Knight draws sword and shield
Troll swings sword.